// src/types/mock-data.type.ts
export type MockData = {
  titles: string[];
  descriptions: string[];
  postDates: string[];
  genres: string[];
  releaseDates: string[];
  ratings: string[];
  previewVideos: string[];
  videos: string[];
  actors: string[];
  directors: string[];
  runTimes: string[];
  commentsCounts: string[];
  posters: string[];
  users: string[];
  emails: string[];
  avatars: string[];
  passwords: string[];
  backgroundImgs: string[];
  backgroundColors: string[];
};